Add event scheduler and world integration
