+++ Test/Test/test_world_events.cpp
@@ -1,6 +1,7 @@
 #include "../../Game/world.hpp"
 #include "../../Game/event.hpp"
 #include "../../System_utils/test_runner.hpp"
+#include "../../Template/vector.hpp"
 
 FT_TEST(test_world_process_events, "ft_world event expiration")
 {
@@ -8,15 +9,18 @@ FT_TEST(test_world_process_events, "ft_world event expiration")
     ft_event event;
     event.set_id(1);
     event.set_duration(3);
-    world.get_events().insert(event.get_id(), event);
+    world.schedule_event(event);
 
-    world.process_events(1);
-    Pair<int, ft_event>* event_entry = world.get_events().find(1);
-    if (!event_entry || event_entry->value.get_duration() != 2)
+    world.update_events(1);
+    ft_vector<ft_event> events;
+    world.get_event_scheduler().dump_events(events);
+    if (events.size() != 1 || events[0].get_duration() != 2)
         return (0);
 
-    world.process_events(2);
-    if (world.get_events().find(1))
+    world.update_events(2);
+    events.clear();
+    world.get_event_scheduler().dump_events(events);
+    if (events.size() != 0)
         return (0);
 
     return (1);
